Using concrete data for a more predictable flow of work
Wouldn’t it be nice if you could use an agile approach and have good data for forecasting when something was going to be done, could shorten cycle time, optimize predictability and delivery speed? I’ll bet it would!

This presentation addresses these questions by using a Kanban with Scrum approach, focusing on metrics around a flow-driven work effort for predictability. By the end of the session you’ll have a strategy for collecting and understanding data that can be applied to any of the above questions.